requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
pydantic-settings>=2.2.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0  # passlib 1.7.4 is incompatible with bcrypt>=4.0
httpx>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
itsdangerous>=2.1.0
starlette>=0.36.0

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
